+++ src/rules/prototypes.jl
@@ -155,10 +155,6 @@ include("implication/in2.jl")
 include("implication/out.jl")
 include("implication/marginals.jl")
 
-# include("delta/helpers.jl")
-
-include("delta/helpers/shared.jl")
-
 include("delta/linearization/in.jl")
 include("delta/linearization/out.jl")
 include("delta/linearization/marginals.jl")
